long long support
    patch from
    https://lists.gnu.org/archive/html/tinycc-devel/2015-05/txtRZFLRPDau1.txt
    --------------------------------------------------------------------------
    This patch set lets cfront 3.0.3 support the "long long" data type and the
    corresponding constants, which makes it interoperable with c99 toolchains.

    Note that the copyright notice and the license below cover only my
    changes, not to be confused with the terms governing the use and
    distribution of cfront.
    --------------------------------------------------------------------------
    Copyright (c) 2014, RL @ Aetey Global Technologies AB
    All rights reserved.

    Redistribution and use in source and binary forms, with or without
    modification, are permitted provided that the following conditions
    are met:

    1. Redistributions of source code must retain the above copyright notice,
    this list of conditions and the following disclaimer.

    2. Redistributions in binary form must reproduce the above copyright
    notice, this list of conditions and the following disclaimer in the
    documentation and/or other materials provided with the distribution.

    THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
    "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
    LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
    A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
    HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
    SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
    TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
    PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
    LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
    NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
    SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

+++ src/print.C
@@ -704,7 +704,7 @@ void expr::print()
 	{	// look for (&a == 0) etc.
 		Neval = 0;
 		binary_val = 1;
-		long i = cond->eval();
+		long long i = cond->eval();
 		binary_val = 0;
 		if (Neval == 0)
 			(i?e1:e2)->print();
